feat(aspire): bootstrap Azurite invoices container with public-blob access
Azurite ships empty (no containers, no CORS, no public access). The first
browser upload to uploadScan.ts was 404ing with ContainerNotFound, and
thumbnail GETs were 403ing because the default access policy blocks
anonymous reads.

Extend the existing CORS bootstrap to also CreateIfNotExistsAsync each
container with PublicAccessType.Blob. For pre-existing containers (where
CreateIfNotExistsAsync is a no-op and ignores the access policy arg),
explicitly upgrade the ACL via SetAccessPolicyAsync — covers the path
where the data volume persists a container created by an older bootstrap
that didn't set public-blob access.

Renamed AzuriteCorsBootstrap -> AzuriteBootstrap to reflect the broader
scope (CORS + container provisioning).
